+++ packages/core/types/types/index.d.ts
@@ -1,4 +1,4 @@
 export { RegisterInteraction, GetInteraction, Interaction, Interactions } from "./interaction";
-export { Route, RouteDescriptor, ParamParser, ParamParsers, MatchedRouteProps, ResponseBuilder, EveryMatchFn, InitialMatchFn, OnFns } from "./route";
-export { Response, RawParams, Params, ModifiableResponseProperties } from "./response";
+export { Route, RouteDescriptor, ParamParser, ParamParsers, ResponseBuilder, EveryMatchFn, InitialMatchFn, OnFns } from "./route";
+export { Response, RawParams, Params, MatchResponseProperties, SettableResponseProperties } from "./response";
 export { CuriRouter, RouterOptions, ResponseHandler, Emitted, RemoveResponseHandler, SideEffect, Cache, Navigation } from "./curi";