+++ packages/core/types/types/match.d.ts
@@ -1,15 +1,15 @@
 import { InternalRoute } from "./route";
-import { Params, GenericResponse } from "./response";
+import { Params, MatchResponseProperties } from "./response";
 export interface MatchingRoute {
     route: InternalRoute;
     params: Params;
 }
 export interface MissMatch {
     route: undefined;
-    response: undefined;
+    match: undefined;
 }
 export interface Match {
     route: InternalRoute;
-    response: GenericResponse;
+    match: MatchResponseProperties;
 }
 export declare type PossibleMatch = Match | MissMatch;